==> p_15.py <==
import numpy as np
import matplotlib.pyplot as plt
from sklearn import svm, preprocessing
import pandas as pd
from matplotlib import style
import os
style.use("ggplot")

FEATURES = ['DE Ratio', 
            'Trailing P/E', 
            'Price/Sales', 
            'Price/Book', 
            'Profit Margin', 
            'Operating Margin', 
            'Return on Assets', 
            'Return on Equity', 
            'Revenue Per Share', 
            'Market Cap', 
            'Enterprise Value', 
            'Forward P/E', 
            'PEG Ratio', 
            'Enterprise Value/Revenue', 
            'Enterprise Value/EBITDA', 
            'Revenue', 
            'Gross Profit', 
            'EBITDA', 
            'Net Income Avl to Common ', 
            'Diluted EPS', 
            'Earnings Growth', 
            'Revenue Growth', 
            'Total Cash', 
            'Total Cash Per Share', 
            'Total Debt', 
            'Current Ratio', 
            'Book Value Per Share', 
            'Cash Flow', 
            'Beta', 
            'Held by Insiders', 
            'Held by Institutions', 
            'Shares Short (as of', 
            'Short Ratio', 
            'Short % of Float', 
            'Shares Short (prior ']

def Build_Data_Set(features = ["DE Ratio", 
                               "Trailing P/E"]):
    data_df = pd.DataFrame.from_csv(os.path.join("C:/Users/Aman Pahariya/Desktop/aman_doc/ML","key_stats.csv"))
    #data_df = data_df[:100]
    data_df = data_df.reindex(np.random.permutation(data_df.index))
    X = np.array(data_df[FEATURES].values)
    y = (data_df["Status"].replace("underperform",0).replace("outperform",1).values.tolist())
    X = preprocessing.scale(X)
    
    return X,y

def Analysis():
    X,y = Build_Data_Set()
    test_size = 1000
    print(len(X))
    clf = svm.SVC(kernel = "linear" , C = 1.0)
    clf.fit(X,y)
    correct_count = 0
    for x in range(1, test_size+1):
        if clf.predict(X[-x])[0] == y[-x]:
            correct_count += 1
            #print(correct_count)
    
    
    print("Accuracy:" , (correct_count*100.00/test_size))
    
    
def randomizing():
    df = pd.DataFrame({"D1":range(5), "D2":range(5)})
    print(df)
    df2 = df.reindex(np.random.permutation(df.index))
    print(df2)
    
#randomizing()
Analysis()
